+++ src/Core/DevicePreview/InvalidDeviceIDError.ts
@@ -1,0 +1,5 @@
+export class InvalidDeviceIDError extends Error {
+  constructor(id: string) {
+    super(`${id} is invalid`)
+  }
+}